requirements.txt
behave==1.2.6
selenium==4.18.1
allure-behave==2.13.5
webdriver-manager==4.0.1
python-dotenv==1.0.1
